+++ src/js/main.js
@@ -1,8 +1,8 @@
-import Slider from './modules/slider';
+import MainSlider from './modules/slider/slider-main';
 import VideoPlayer from './modules/playVideo';
 
 window.addEventListener('DOMContentLoaded', () => {
-	const slider = new Slider('.page', '.next');
+	const slider = new MainSlider({ page: '.page', btns: '.next' });
 	slider.render();
 
 	const player = new VideoPlayer('.showup .play', '.overlay');
