+++ src/js/modules/slider/slider-main.js
@@ -1,10 +1,10 @@
-export default class Slider {
-	constructor(page, btns) {
-		this.page = document.querySelector(page);
-		this.slides = this.page.children;
-		this.btns = document.querySelectorAll(btns);
-		this.slideIndex = 1;
+import Slider from './slider';
+
+export default class MainSlider extends Slider {
+	constructor (page, btns) {
+		super(page,btns);
 	}
+
 	showSlides(n) {
 		if (n > this.slides.length) {
 			this.slideIndex = 1;
@@ -52,7 +52,7 @@ export default class Slider {
 				e.preventDefault();
 				this.slideIndex = 1;
 				this.showSlides(this.slideIndex);
-			})
+			});
 		});
 
 		this.showSlides(this.slideIndex);
